Fix logic error in vApp#power_on test
Whilst fixing up lint errors, I noticed that this test
started to fail when fixing the '==' check to use eq()

It appears that expect(blah) == true is always true!

The test logic was wrong -- returning :status => 8 each
time would emulate a machine NOT powering on.

Updating the test to return 8 (STOPPED), then 4 (RUNNING)
fixes the test.