Optionally cause Rspec tests to fail fast
Optionally cause Rspec tests to fail on the first error encountered.

The vCloud Tools don't handle ^C well on my machine so it's not easy to
stop tests once they have started. Given that integration tests can be
slow to complete, I think it's useful for them to fail on the first
error.

Use the `VCLOUD_TOOLS_RSPEC_FAIL_FAST` environment variable to determine
whether Rspec should fail fast or not. The default is leave `fail_fast`
disabled, as was this case before this commit.

More about `fail_fast`:
https://www.relishapp.com/rspec/rspec-core/v/2-14/docs/configuration/fail-fast