Fail integration tests fast
Cause integration tests to fail on the first error encountered.

The vCloud Tools don't handle ^C well so it's not easy to stop tests
once they have started. Given that integration tests can be slow to
complete, I think it's useful for them to fail on the first error.

More about this option:
https://www.relishapp.com/rspec/rspec-core/v/2-14/docs/configuration/fail-fast