Basic Vm#attach_independent_disk integration test
At present there is no way of telling that an independent disk has been
attached to a VM successfully.

For now, assume that the attach_independent_disks method returning true is
enough of a measure of success (which at least covers the case where Fog
raises an exception, or the process_task call fails).

In a later commit, a method of establishing the vms that a disk is attached to
will be added, and this test improved.

UPDATE: Noticed later on that get_vapp(vm_id) does in fact report that independent
        disks are attached. A RASD OVF disk item will be present containing a
        ':vcloud_disk' key, with the link to the disk as its value.