Add IndependentDisk#attached_vms method
... to allow us to know where a Disk has been attached to, so we can check
that it has been attached correctly!

(NB: There is no way of doing this directly from the get_vm / Vm#vcloud_attributes
call, as we can for local disks)

This is mainly being added so that we can be sure the Vm#attach_independent_disks
test is working more reliably, but it is a useful addition to the IndependentDisk class.