Improve attach disk test; note ordering issue
Improved the VM attach independent disk tests, to check that
the disk was actually flagged as attached to the VM via the
IndependentDisk#attached_vms call.

In the process of this, discovered that the Vm#add_extra_disks method
fails if there are independent disks already attached to the VM. I
strongly suspect that this is due to the way the Fog Model handles
adding local disks rather than it being a limitation of the vCD API itself,
but fixing that is out of scope for this update.

Added a test to document this behaviour and the exception raised by Fog.