Do not allow creation of duplicate disks
The vCloud Director API allows you to create a rather nasty situation with
independent disks, as it does not enforce uniqueness on independent disk name
- even when they are in the same vDC. Once duplicate disk names have been created
in the same vDC, it is very difficult to tell them apart.

This commit adds code to IndependentDisk.create to check for a disk with the same name+vdc
combination before creation, and raise an error if it is attempted.

In the process, I have added specific exception classes for the various error cases of
the get_by_name_and_vdc_name() method.