Allow plaintext passwords
While banning the use of plaintext passwords is a noble goal,
currently there is no other non-interactive way to authenticate
with Skyscape (and, I assume, other vCloud-based providers). We
use this to allow us to add vcloud-tools to our CI.

The fog file should be locked down to mitigate the vulnerability
this feature exposes.